refactor(bot-client): extract truncationGate primitives to shared utils
Move detection, embed builders, and button-row builders from
character/truncationWarning.ts into utils/dashboard/truncationGate/ so
both character and persona dashboards (and future ones) can share them.
Detection function is generic over T; button builders take entityType
as first parameter to produce per-dashboard custom IDs.

Pure refactor — zero behavior change for character. Existing
truncationWarning.test.ts loses its detection/embed/button-shape
describes (relocated to truncationGate/*.test.ts) and keeps the four
handler describes plus showTruncationWarning.
